expression: wrap with convert_charset() in utf8 strings
